renaming video_to_npy.py to load_data_prediction.py to avoid confusion, renaming pre_processing.py to pre_processing_prediction.py because pre_processing module will only be used in prediction phase, test all functions with frontend.py, all working good
